Generate debian/control from debian/control.in
There are versions of Python which are available in official repository
for Ubuntu, but not available in official repositories for Debian
Unstable. And vice versa.

For building Debian and Ubuntu packages from ithe same source it's neccessary
to generate different debian/control files for each of these
distributions with specific set of uwsgi-pythonX.X packages.