Use alternatives system for providing default python plugin
Binary packages uwsgi-plugin-python/uwsgi-plugin-python3 are providing
appropriate python_plugin.so/python3_plugin.so with using of
alternatives system.

pythonX_plugin.so is symlinked to pythonX.Y_plugin.so, where X.Y is the
version of default Python runtime.

Also hooks for update to installed Python runtime are provided by
uwsgi-plugin-python/uwsgi-plugin-python3 packages. With updating of
default Python runtime, link to underlying alternative of
pythonX_plugin.so is also updated.